requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
Pillow>=10.0.0
python-dotenv>=1.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=2024.0
django-filter>=23.0
mssql-django>=1.3

# AI Dependencies - Tất cả đều miễn phí!
# Ollama: Chạy local (không cần package)
# OpenRouter: requests (đã có sẵn)
# Groq: requests (đã có sẵn)
requests>=2.31.0
google-generativeai>=0.3.0

# Optional - cho vector search (nếu cần)
# chromadb>=0.4.0
# sentence-transformers>=2.2.0
numpy>=1.24.0
